pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "slides-agent"
version = "0.1.0"
description = "AI-first CLI for Google Slides — deterministic, JSON-native, agent-optimized"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "slides-agent contributors" }]
keywords = ["google-slides", "cli", "ai", "agent", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
]

dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "rich>=13.0",
    "httplib2>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
slides-agent = "slides_agent.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"
